[YS-151] refactor: 페이지단의 스타일 코드 분리
